+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/ChartWidgets/messages.ts
@@ -1,0 +1,45 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  analyticsChart: {
+    id: 'app.containers.admin.ReportBuilder.charts.analyticsChart',
+    defaultMessage: 'Chart',
+  },
+  analyticsChartTitle: {
+    id: 'app.containers.admin.ReportBuilder.charts.analyticsChartTitle',
+    defaultMessage: 'Title',
+  },
+  analyticsChartDateRange: {
+    id: 'app.containers.admin.ReportBuilder.charts.analyticsChartDateRange',
+    defaultMessage: 'Date range',
+  },
+  visitorTimeline: {
+    id: 'app.containers.admin.ReportBuilder.charts.visitorTimeline',
+    defaultMessage: 'Visitor timeline',
+  },
+  activeUsersTimeline: {
+    id: 'app.containers.admin.ReportBuilder.charts.activeUsersTimeline',
+    defaultMessage: 'Participants timeline',
+  },
+  totalParticipants: {
+    id: 'app.modules.commercial.analytics.admin.components.ActiveUsersCard.totalParticipants',
+    defaultMessage: 'Total participants',
+  },
+  trafficSources: {
+    id: 'app.containers.admin.ReportBuilder.charts.trafficSources',
+    defaultMessage: 'Traffic sources',
+  },
+  usersByGender: {
+    id: 'app.containers.admin.ReportBuilder.charts.usersByGender',
+    defaultMessage: 'Users by gender',
+  },
+  usersByAge: {
+    id: 'app.containers.admin.ReportBuilder.charts.usersByAge',
+    defaultMessage: 'Users by age',
+  },
+  noData: {
+    id: 'app.containers.admin.ReportBuilder.charts.noData',
+    defaultMessage:
+      'There is no data available for the filters you have selected.',
+  },
+});