+++ front/app/modules/commercial/analytics/admin/components/EmailDeliveriesCard/Chart/index.tsx
@@ -17,9 +17,9 @@ import { toThreeLetterMonth } from 'utils/dateUtils';
 import { generateEmptyData } from './generateEmptyData';
 
 // typings
-import { Dates, Resolution } from '../../../typings';
+import { Dates, Resolution } from 'components/admin/GraphCards/typings';
 import { LegendItem } from 'components/admin/Graphs/_components/Legend/typings';
-import { TimeSeries } from '../../../hooks/useEmailDeliveries/typings';
+import { TimeSeries } from '../useEmailDeliveries/typings';
 
 type Props = Dates &
   Resolution & {